+++ packages/notifi-dapp-example/src/components/NotifiSignUpButton.tsx
@@ -136,14 +136,12 @@ export const NotifiSignUpButton: React.FC<NotifiSignUpButtonProps> = ({
   }, [email, phoneNumber, telegramId, useDiscord, data.isContactInfoRequired]);
 
   return (
-    <>
-      <button
-        className="rounded-lg bg-notifi-button-primary-blueish-bg text-notifi-button-primary-text w-72 h-11 mb-6 text-sm font-bold disabled:hover:bg-notifi-button-primary-blueish-bg hover:bg-notifi-button-hover-bg"
-        disabled={!isInitialized || loading || hasErrors || !isInputFieldsValid}
-        onClick={onClick}
-      >
-        <span>{loading ? 'Loading' : buttonText}</span>
-      </button>
-    </>
+    <button
+      className="rounded-lg bg-notifi-button-primary-blueish-bg text-notifi-button-primary-text w-72 h-11 mb-6 text-sm font-bold disabled:hover:bg-notifi-button-primary-blueish-bg hover:bg-notifi-button-hover-bg"
+      disabled={!isInitialized || loading || hasErrors || !isInputFieldsValid}
+      onClick={onClick}
+    >
+      <span>{loading ? 'Loading' : buttonText}</span>
+    </button>
   );
 };
